Factor out the islands map generators
This breaks up the very large mapgen.cpp further, almost bringing it
down to a manageable size.
