+++ server/generator/islands.h
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: GPLv3-or-later
+// SPDX-FileCopyrightText: Freeciv21 authors
+// SPDX-FileCopyrightText: Freeciv authors
+// SPDX-FileCopyrightText: Louis Moureaux <m_louis30@yahoo.com>
+
+void mapgenerator2();
+void mapgenerator3();
+void mapgenerator4();
+
+void island_terrain_init();
+void island_terrain_free();
