+++ server/generator/mapgen.h
@@ -15,6 +15,18 @@
 #define HAS_POLES                                                           \
   (wld.map.server.temperature < 70 && !wld.map.server.alltemperate)
 
+/**
+   Conditions used mainly in rand_map_pos_characteristic()
+ */
+// WETNESS
+
+// necessary condition of deserts placement
+#define map_pos_is_dry(ptile)                                               \
+  (map_colatitude((ptile)) <= DRY_MAX_LEVEL                                 \
+   && map_colatitude((ptile)) > DRY_MIN_LEVEL                               \
+   && count_terrain_class_near_tile((ptile), false, true, TC_OCEAN) <= 35)
+typedef enum { WC_ALL = 200, WC_DRY, WC_NDRY } wetness_c;
+
 extern int forest_pct;
 extern int desert_pct;
 extern int swamp_pct;
@@ -25,5 +37,9 @@ extern int river_pct;
 extern struct extra_type *river_types[MAX_ROAD_TYPES];
 extern int river_type_count;
 
+void make_plains();
 void make_polar();
+
+bool test_wetness(const struct tile *ptile, wetness_c c);
+
 bool map_fractal_generate(bool autosize, struct unit_type *initial_unit);
